fix nodemon always restarting bug due to creating new json file every time

+++ common/src/metropolitanLocations.ts
@@ -1,6 +1,7 @@
 /* Put all the metropolitan locations that we're going to use to loadtest here */
+import deepEqual from 'deep-equal'
 import { getHaversineDistanceMiles } from './haversine'
-import { writeFileSync } from 'fs'
+import {readFileSync, writeFileSync, existsSync } from 'fs'
 export const metropolitanLocations = [
   {
     name: "New York City",
@@ -44,7 +45,12 @@ export const metropolitanLocations = [
 // this is for the loadtesting script. We need to create a json that can be read by the load testing script.
 // this is in the server/bin/common/src/ folder
 const metropolitanJsonPath = __dirname + '/../../../src/loadtest/metropolitanLocations.json'
-writeFileSync(metropolitanJsonPath, JSON.stringify(metropolitanLocations));
+
+if (!existsSync(metropolitanJsonPath) || !deepEqual(JSON.parse(readFileSync(metropolitanJsonPath, { encoding: 'utf8'})), metropolitanLocations)) {
+  // create new one
+  writeFileSync(metropolitanJsonPath, JSON.stringify(metropolitanLocations));
+  console.log("Updated src/loadtest/metropolitanLocations.json")
+}
 
 export const getNearestMetroLocation = (lat: number, long: number) => {
   const metroAreaDistances = metropolitanLocations
